fix: entityManagerFactory에러 수정 완료

+++ src/main/java/com/aidredaline/backend/domain/template/dto/TemplateResponse.java
@@ -16,7 +16,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class TemplateResponse {
 
-    private Long templateId;
+    private Integer templateId;
     private String name;
     private String category;
     private String svgPath;
